test(FSADT1-1365): updated test
