+++ legacy/src/test/java/ca/bc/gov/app/service/ClientSearchServiceIntegrationTest.java
@@ -3,14 +3,14 @@ package ca.bc.gov.app.service;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
+import ca.bc.gov.app.dto.AddressSearchDto;
 import ca.bc.gov.app.dto.ForestClientDto;
 import ca.bc.gov.app.exception.MissingRequiredParameterException;
 import ca.bc.gov.app.extensions.AbstractTestContainerIntegrationTest;
-import java.time.LocalDate;
+import java.util.List;
 import java.util.stream.Stream;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -18,138 +18,83 @@ import org.springframework.beans.factory.annotation.Autowired;
 import reactor.test.StepVerifier;
 import reactor.test.StepVerifier.FirstStep;
 
-@DisplayName("Integrated Test | Client Search Service-")
+@DisplayName("Integrated Test | Client Search Service")
 class ClientSearchServiceIntegrationTest extends AbstractTestContainerIntegrationTest {
 
   @Autowired
   private ClientSearchService service;
 
-  @DisplayName("should find by registration number or company name")
-  @ParameterizedTest
-  @MethodSource("byNumberOrCompanyName")
-  void shouldFindByRegistrationNumberOrCompanyName(
-      String registrationNumber,
-      String companyName,
-      String expected,
-      Class<RuntimeException> exception
-  ) {
-    StepVerifier.FirstStep<ForestClientDto> test =
-        service
-            .findByRegistrationNumberOrCompanyName(registrationNumber, companyName)
-            .as(StepVerifier::create);
-
-    verifyTestData(expected, exception, test);
-  }
-
-  @DisplayName("should find individual")
-  @ParameterizedTest
-  @MethodSource("byIndividual")
-  void shouldFindIndividual(
-      String firstName,
-      String lastName,
-      LocalDate dob,
-      String identification,
-      String expected,
-      Class<RuntimeException> exception
-  ) {
-    StepVerifier.FirstStep<ForestClientDto> test =
-        service
-            .findByIndividual(firstName, lastName, dob, identification, true)
-            .as(StepVerifier::create);
-
-    verifyTestData(expected, exception, test);
-
-  }
-
-  @DisplayName("should find by match")
-  @ParameterizedTest
-  @MethodSource("byMatch")
-  void shouldFindByMatch(
-      String companyName,
-      String expected,
-      Class<RuntimeException> exception
-  ) {
-    StepVerifier.FirstStep<ForestClientDto> test =
-        service
-            .matchBy(companyName)
-            .as(StepVerifier::create);
-
-    verifyTestData(expected, exception, test);
-
-  }
-
-  @DisplayName("should find by id and last name")
-  @ParameterizedTest
-  @MethodSource("byIdLastName")
-  void shouldFindByLastNameAndDocument(
-      String lastName,
-      String identification,
-      String expected,
-      Class<RuntimeException> exception
-  ) {
-    StepVerifier.FirstStep<ForestClientDto> test =
-        service
-            .findByIdAndLastName(identification, lastName)
-            .as(StepVerifier::create);
-
-    verifyTestData(expected, exception, test);
-
-  }
-
-  @DisplayName("should find by document")
-  @ParameterizedTest
-  @MethodSource("byDocument")
-  void shouldFindByIdLas(
-      String idType,
-      String identification,
-      String expected,
-      Class<RuntimeException> exception
-  ) {
-    StepVerifier.FirstStep<ForestClientDto> test =
-        service
-            .findByIdentification(idType, identification)
-            .as(StepVerifier::create);
-
-    verifyTestData(expected, exception, test);
-
-  }
-
   @DisplayName("should find by email on contact and location")
   @ParameterizedTest
-  @MethodSource("byDocument")
-  void shouldfindByGeneralEmail() {
+  @MethodSource("byEmail")
+  void shouldFindByGeneralEmail(
+      String email,
+      List<String> expected,
+      Class<RuntimeException> exception
+  ) {
+
+    FirstStep<ForestClientDto> test =
+        service
+            .findByGeneralEmail(email)
+            .as(StepVerifier::create);
+
+    verifyTestData(expected, exception, test);
 
   }
 
-  @DisplayName("should find by email on contact and location")
-  @Test
-  void shouldFindByAddress() {
-    StepVerifier.FirstStep<ForestClientDto> test =
-    service
-        .findByEntireAddress(
-            "510 FULTON PLAZA",
-            "FORT MCMURRAY",
-            "AB",
-            "T9J9R1",
-            "CANADA"
-        )
-        .as(StepVerifier::create);
+  @DisplayName("should find by phone on contact and location")
+  @ParameterizedTest
+  @MethodSource("byPhone")
+  void shouldFindByGeneralPhone(
+      String phone,
+      List<String> expected,
+      Class<RuntimeException> exception
+  ) {
 
-    verifyTestData("00000123", null, test);
+    FirstStep<ForestClientDto> test =
+        service
+            .findByGeneralPhoneNumber(phone)
+            .as(StepVerifier::create);
+
+    verifyTestData(expected, exception, test);
+
+  }
+
+  @DisplayName("should find by location")
+  @ParameterizedTest
+  @MethodSource("byLocation")
+  void shouldFindByLocation(
+      AddressSearchDto location,
+      String expected,
+      Class<RuntimeException> exception
+  ) {
+
+    FirstStep<ForestClientDto> test =
+        service
+            .findByEntireAddress(location)
+            .as(StepVerifier::create);
+
+    verifyTestData(expected, exception, test);
+
   }
 
   private void verifyTestData(
-      String expected,
+      List<String> expectedList,
       Class<RuntimeException> exception,
       FirstStep<ForestClientDto> test
   ) {
-    if (StringUtils.isNotBlank(expected)) {
-      test
-          .assertNext(dto -> {
-            assertNotNull(dto);
-            assertEquals(expected, dto.clientNumber());
-          });
+    if (expectedList != null && !expectedList.isEmpty()) {
+      for(String expected: expectedList) {
+        if (StringUtils.isNotBlank(expected)) {
+          test
+              .assertNext(dto -> {
+                assertNotNull(dto);
+                assertEquals(expected, dto.clientNumber());
+              });
+        }
+      }
     }
+
 
     if (exception != null) {
       test.expectError(exception);
@@ -158,69 +103,77 @@ class ClientSearchServiceIntegrationTest extends AbstractTestContainerIntegratio
     test.verifyComplete();
   }
 
-  private static Stream<Arguments> byNumberOrCompanyName() {
-    return Stream.of(
-        Arguments.of("BC123444789", null, "00000011", null),
-        Arguments.of(null, "INDIAN CANADA", "00000006", null),
-        Arguments.of(null, "ELARICHO", "00000005", null),
-        Arguments.of("AB994455454", null, StringUtils.EMPTY, null),
-        Arguments.of(null, "Zamasu Corp", StringUtils.EMPTY, null),
-        Arguments.of(null, null, null, MissingRequiredParameterException.class)
+  private void verifyTestData(
+      String expected,
+      Class<RuntimeException> exception,
+      FirstStep<ForestClientDto> test
+  ) {
+    verifyTestData(
+        expected == null ? List.of() : List.of(expected),
+        exception,
+        test
     );
   }
 
-  private static Stream<Arguments> byIndividual() {
-    return Stream
-        .of(
-            Arguments.of("JAMES", "BAXTER", LocalDate.of(1959, 5, 18), null, "00000001", null),
-            Arguments.of("THOMAS", "FUNNY", LocalDate.of(1939, 7, 4), "34458787", "00000002", null),
-            Arguments.of("ALBUS", "DUMBLEDORE", LocalDate.of(1814, 5, 12), null, StringUtils.EMPTY,
-                null),
-            Arguments.of("JAMES", null, null, null, null, MissingRequiredParameterException.class),
-            Arguments.of("JAMES", "Baxter", null, null, null,
-                MissingRequiredParameterException.class),
-            Arguments.of(null, "Baxter", LocalDate.of(1959, 5, 18), null, null,
-                MissingRequiredParameterException.class),
-            Arguments.of(null, null, LocalDate.of(1959, 5, 18), null, null,
-                MissingRequiredParameterException.class)
+  private static Stream<Arguments> byEmail() {
+    return
+        Stream.concat(
+            emptyCases(),
+            Stream
+                .of(
+                    Arguments.of("celinedion@email.ca", List.of(), null),
+                    Arguments.of("uturfes0@cnn.com", List.of("00000103"), null),
+                    Arguments.of("mail@mail.ca", List.of("00000006","00000001","00000003"), null)
+                )
         );
   }
 
-  private static Stream<Arguments> byMatch() {
-    return Stream
-        .of(
-            Arguments.of("BAXTER", "00000001", null),
-            Arguments.of("SAMPLE INDIAN BAND COUNC", "00000004", null),
-            Arguments.of("FALCONI CORP", StringUtils.EMPTY, null),
-            Arguments.of(StringUtils.EMPTY, StringUtils.EMPTY,
-                MissingRequiredParameterException.class),
-            Arguments.of(null, StringUtils.EMPTY, MissingRequiredParameterException.class)
+  private static Stream<Arguments> byPhone() {
+    return
+        Stream.concat(
+            emptyCases(), Stream
+                .of(
+                    Arguments.of("1232504567", List.of(), null),
+                    Arguments.of("2502502550", List.of("00000004","00000003","00000001"), null),
+                    Arguments.of("2894837433", List.of("00000103"), null)
+                )
         );
   }
 
-  private static Stream<Arguments> byIdLastName() {
+  private static Stream<Arguments> byLocation() {
     return Stream
         .of(
-            Arguments.of("FUNNY", "34458787", "00000002", null),
-            Arguments.of("Baxter", null, null,
-                MissingRequiredParameterException.class),
-            Arguments.of(null, null, null,
-                MissingRequiredParameterException.class)
+            Arguments.of(new AddressSearchDto(
+                "",
+                "",
+                "",
+                "",
+                ""
+            ), null, MissingRequiredParameterException.class),
+            Arguments.of(new AddressSearchDto(
+                "510 FULTON PLAZA",
+                "FORT MCMURRAY",
+                "AB",
+                "T9J9R1",
+                ""
+            ), null, MissingRequiredParameterException.class),
+            Arguments.of(new AddressSearchDto(
+                "510 FULTON PLAZA",
+                "FORT MCMURRAY",
+                "AB",
+                "T9J9R1",
+                "CANADA"
+            ), "00000123", null)
         );
   }
 
-  private static Stream<Arguments> byDocument() {
+  private static Stream<Arguments> emptyCases() {
     return Stream
         .of(
-            Arguments.of("BCDL", "9994457", "00000005", null),
-            Arguments.of("BCSC", "Wull", "00000007", null),
-            Arguments.of("BCDL", "3334447", StringUtils.EMPTY, null),
-            Arguments.of("BCDL", null, null,
-                MissingRequiredParameterException.class),
-            Arguments.of(null, "9994457", null,
-                MissingRequiredParameterException.class),
-            Arguments.of(null, null, null,
-                MissingRequiredParameterException.class)
+            Arguments.of(null, null, MissingRequiredParameterException.class),
+            Arguments.of(StringUtils.EMPTY, null, MissingRequiredParameterException.class),
+            Arguments.of("  ", null, MissingRequiredParameterException.class)
         );
   }
+
 }
